+++ foxes/input/states/field_data.py
@@ -1,7 +1,9 @@
 import numpy as np
+import matplotlib.pyplot as plt
 
 from foxes.utils import weibull_weights
-from foxes.config import config
+from foxes.config import config, get_output_path
+from foxes.output import FarmLayoutOutput
 import foxes.variables as FV
 import foxes.constants as FC
 
@@ -25,11 +27,9 @@ class FieldData(DatasetStates):
         The height coordinate name in the data
     weight_ncvar: str
         Name of the weight data variable in the nc file(s)
-    bounds_extra_space: float or str
-        The extra space, either float in m,
-        or str for units of D, e.g. '2.5D'
-    height_bounds: tuple, optional
-        The (h_min, h_max) height bounds in m. Defaults to H +/- 0.5*D
+    grid_point_plot: str, optional
+        Path to a plot file, e.g. grid_points.png, to visualize the
+        selected data grid points and the layout of the farm.
 
     Examples
     --------
@@ -58,8 +58,7 @@ class FieldData(DatasetStates):
         h_coord="height",
         time_format=r"%Y-%m-%d_%H:%M:%S",
         weight_ncvar=None,
-        bounds_extra_space=1000,
-        height_bounds=None,
+        grid_point_plot=None,
         **kwargs,
     ):
         """
@@ -81,11 +80,9 @@ class FieldData(DatasetStates):
             The datetime parsing format string
         weight_ncvar: str, optional
             Name of the weight data variable in the nc file(s)
-        bounds_extra_space: float or str, optional
-            The extra space, either float in m,
-            or str for units of D, e.g. '2.5D'
-        height_bounds: tuple, optional
-            The (h_min, h_max) height bounds in m. Defaults to H +/- 0.5*D
+        grid_point_plot: str, optional
+            Path to a plot file, e.g. grid_points.png, to visualize the
+            selected data grid points and the layout of the farm.
         kwargs: dict, optional
             Additional parameters for the base class
 
@@ -96,8 +93,7 @@ class FieldData(DatasetStates):
         self.y_coord = y_coord
         self.h_coord = h_coord
         self.weight_ncvar = weight_ncvar
-        self.bounds_extra_space = bounds_extra_space
-        self.height_bounds = height_bounds
+        self.grid_point_plot = grid_point_plot
 
         assert FV.WEIGHT not in self.ovars, (
             f"States '{self.name}': Cannot have '{FV.WEIGHT}' as output variable, got {self.ovars}"
@@ -119,37 +115,82 @@ class FieldData(DatasetStates):
         if self.h_coord is not None:
             self._cmap[FV.H] = self.h_coord
 
-    def load_data(self, algo, verbosity=0):
+    def preproc_first(
+        self,
+        algo,
+        data,
+        bounds_extra_space,
+        height_bounds,
+        verbosity=0,
+    ):
         """
-        Load and/or create all model data that is subject to chunking.
-
-        Such data should not be stored under self, for memory reasons. The
-        data returned here will automatically be chunked and then provided
-        as part of the mdata object during calculations.
+        Preprocesses the first file.
 
         Parameters
         ----------
         algo: foxes.core.Algorithm
             The calculation algorithm
+        data: xarray.Dataset
+            The dataset to preprocess
+        bounds_extra_space: float or str, optional
+            The extra space, either float in m,
+            or str for units of D, e.g. '2.5D'
+        height_bounds: tuple, optional
+            The (h_min, h_max) height bounds in m. Defaults to H +/-
         verbosity: int
             The verbosity level, 0 = silent
 
-        Returns
-        -------
-        idata: dict
-            The dict has exactly two entries: `data_vars`,
-            a dict with entries `name_str -> (dim_tuple, data_ndarray)`;
-            and `coords`, a dict with entries `dim_name_str -> dim_array`
-
         """
-        return super().load_data(
+
+        super().preproc_first(
             algo,
-            cmap=self._cmap,
-            variables=self.variables,
-            bounds_extra_space=self.bounds_extra_space,
-            height_bounds=self.height_bounds,
+            data,
+            bounds_extra_space=bounds_extra_space,
+            height_bounds=height_bounds,
             verbosity=verbosity,
         )
+
+        if self.grid_point_plot is not None:
+            try:
+                if self.isel is not None:
+                    data = data.isel(self.isel)
+                if self.sel is not None:
+                    data = data.sel(self.sel)
+                has_data = True
+                for c, s in data.sizes.items():
+                    if s == 0:
+                        has_data = False
+                        break
+            except KeyError:
+                has_data = False
+
+            if has_data:
+                fpath = get_output_path(self.grid_point_plot)
+                if verbosity > 0:
+                    print(f"States '{self.name}': Writing grid point plot to '{fpath}'")
+                fig, ax = plt.subplots(figsize=(8, 8))
+                xx, yy = np.meshgrid(
+                    data[self._cmap[FV.X]].values.flatten(),
+                    data[self._cmap[FV.Y]].values.flatten(),
+                )
+                ax.plot(
+                    xx,
+                    yy,
+                    c="blue",
+                    alpha=0.2,
+                    marker=".",
+                    linestyle="None",
+                )
+                anno = 3 if len(algo.farm.wind_farm_names) > 1 else 0
+                FarmLayoutOutput(farm=algo.farm).get_figure(
+                    fig=fig, ax=ax, annotate=anno, fontsize=12
+                )
+                ax.set_xlabel(f"{FV.X} [m]")
+                ax.set_ylabel(f"{FV.Y} [m]")
+                ax.set_aspect("equal", adjustable="box")
+                ax.autoscale_view(tight=True)
+                fig.savefig(fpath, bbox_inches="tight")
+                plt.close()
 
 
 class WeibullField(FieldData):
@@ -240,7 +281,7 @@ class WeibullField(FieldData):
     def __repr__(self):
         return f"{type(self).__name__}(n_wd={self._n_wd}, n_ws={self._n_ws})"
 
-    def _read_ds(self, ds, cmap, variables, verbosity=0):
+    def _read_ds(self, ds, cmap=None, verbosity=0):
         """
         Helper function for _get_data, extracts data from the original Dataset.
 
@@ -248,10 +289,8 @@ class WeibullField(FieldData):
         ----------
         ds: xarray.Dataset
             The Dataset to read data from
-        cmap: dict
-            A mapping from foxes variable names to Dataset dimension names
-        variables: list of str
-            The variables to extract from the Dataset
+        cmap: dict, optional
+            A mapping from foxes variable names to Dataset dimension names, if not given self._cmap will be used
         verbosity: int
             The verbosity level, 0 = silent
 
@@ -267,7 +306,7 @@ class WeibullField(FieldData):
 
         """
         # read data, using wd_coord as state coordinate
-        coords, data0 = super()._read_ds(ds, cmap, variables, verbosity)
+        coords, data0 = super()._read_ds(ds, cmap=cmap, verbosity=verbosity)
         wd = coords.pop(FC.STATE)
 
         # replace state by wd coordinate